// setup/arcgis-experience-builder-1.6/ArcGISExperienceBuilder/client/dist/widgets/common/search/src/config.ts
export enum SearchServiceType {
  GeocodeService = 'GeocodeService',
  FeatureService = 'FeatureService'
}
export enum SearchResultView {
  ResultPanel = 'Result_Panel',
  OtherWidgets = 'Other_Widgets'
}

export enum ArrangementStyle {
  Style1 = 'Style1',
  Style2 = 'Style2'
}
// ...
export interface GeocodeConfig {
  configId: string
  label: string
  geocodeURL: string
  outputDataSourceId: string
  useDataSource: UseDataSource
  icon?: any
  hint?: string
  searchServiceType?: SearchServiceType
}

export type IMGeocodeConfig = ImmutableObject<GeocodeConfig>

export type DatasourceConfigItem = LayersConfig & GeocodeConfig

export type IMDatasourceConfigItem = ImmutableObject<DatasourceConfigItem>
export interface config {
  datasourceConfig: DatasourceConfigItem[]
  searchResultView: SearchResultView
  arrangementStyle: ArrangementStyle
  hint: string
  maxSuggestions: number
  isUseCurrentLoation: boolean
  isShowRecentSearches: boolean
  recentSearchesMaxNumber: number
  resultMaxNumber: number
  isAutoSelectFirstResult: boolean
  // link
  linkParam?: IMLinkParam
}

export type IMConfig = ImmutableObject<config>